requirements.txt
requests
psycopg2-binary==2.8.6
django-tailwind==2.2.0
Pillow
pyocr